requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Infrastructure dependencies
chromadb>=0.4.0
redis>=5.0.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0
seaborn>=0.12.0

# PDF generation
reportlab>=4.0.0
fpdf2>=2.7.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security/Crypto
cryptography>=41.0.0